pyproject.toml
[tool.poetry]
name = "moneytracker"
version = "0.1.0"
description = ""
authors = ["ArhipDolya"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
python-decouple = "^3.8"
ruff = "^0.1.7" 
pre-commit = "^3.6.0"
django-ninja = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
